💄 Make team overview table feel less crowded
This commit includes two small style changes:

- put the "- invited" in a new line
- add margin after and below change role form

The first change should make it less likely for the change role form to break onto two lines if a user has a long email address.

The second change improves the look of the change role form by adding a margin between the select-element and the submit button.
The margin-bottom is only added for >=sm because for xs bootstrap adds this by itself for all form-groups.