🐛 Use bleach.linkify instead of urlize
Froide converts urls in messages to clickable links, but not email adresses. Until now this was done by using djangos urlize[1] function and removing mailto links afterwards. However urlize should not be used on text that contains html or email adresses that contain single quotes (see note in [1]). This leads to the problem reported in #508.

This switches to bleach.linkify, which can handle existing html code and has an option to ignore emails.

[1]: https://docs.djangoproject.com/en/4.0/ref/templates/builtins/#urlize

Closes #508