🐛 Clear Document.pdf_file after FoiAttachment redaction
If the pdf_file is set (for example because the document pdf needed to be rewritten and get_writeable_file was called), it will take precedence over the attachments pdf. This means that new redactions applied to the attachment after pdf_file is set are not shown in the document, as the old version from pdf_file is still used for generating the pages etc.